refactor: move ApiError into common crate
