+++ crates/common/api_types/beacon/Cargo.toml
@@ -23,6 +23,7 @@ thiserror.workspace = true
 url.workspace = true
 
 # ream dependencies
+ream-api-types-common.workspace = true
 ream-bls.workspace = true
 ream-consensus-beacon.workspace = true
 ream-consensus-misc.workspace = true
