+++ crates/common/api_types/common/Cargo.toml
@@ -10,5 +10,7 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
+actix-web.workspace = true
 alloy-primitives.workspace = true
 serde.workspace = true
+thiserror.workspace = true
