+++ crates/common/api_types/common/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod error;
 pub mod id;
