+++ crates/rpc/beacon/src/handlers/blob_sidecar.rs
@@ -3,10 +3,8 @@ use actix_web::{
     web::{Data, Path},
 };
 use actix_web_lab::extract::Query;
-use ream_api_types_beacon::{
-    error::ApiError, query::BlobSidecarQuery, responses::BeaconVersionedResponse,
-};
-use ream_api_types_common::id::ID;
+use ream_api_types_beacon::{query::BlobSidecarQuery, responses::BeaconVersionedResponse};
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_consensus_beacon::blob_sidecar::BlobIdentifier;
 use ream_storage::{db::ReamDB, tables::table::Table};
 use tree_hash::TreeHash;
