+++ crates/rpc/beacon/src/handlers/committee.rs
@@ -5,11 +5,10 @@ use actix_web::{
     web::{Data, Path, Query},
 };
 use ream_api_types_beacon::{
-    error::ApiError,
     query::{EpochQuery, IndexQuery, SlotQuery},
     responses::BeaconResponse,
 };
-use ream_api_types_common::id::ID;
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_consensus_misc::{constants::beacon::SLOTS_PER_EPOCH, misc::compute_start_slot_at_epoch};
 use ream_storage::db::ReamDB;
 use serde::Serialize;
