+++ crates/rpc/beacon/src/handlers/duties.rs
@@ -4,10 +4,9 @@ use actix_web::{
 };
 use ream_api_types_beacon::{
     duties::{AttesterDuty, ProposerDuty},
-    error::ApiError,
     responses::DutiesResponse,
 };
-use ream_api_types_common::id::ID;
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_consensus_misc::{constants::beacon::SLOTS_PER_EPOCH, misc::compute_start_slot_at_epoch};
 use ream_storage::db::ReamDB;
