+++ crates/rpc/beacon/src/handlers/header.rs
@@ -4,11 +4,10 @@ use actix_web::{
 };
 use alloy_primitives::B256;
 use ream_api_types_beacon::{
-    error::ApiError,
     query::{ParentRootQuery, SlotQuery},
     responses::BeaconResponse,
 };
-use ream_api_types_common::id::ID;
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_consensus_misc::beacon_block_header::SignedBeaconBlockHeader;
 use ream_storage::{db::ReamDB, tables::table::Table};
 use serde::{Deserialize, Serialize};
