+++ crates/rpc/beacon/src/handlers/prepare_beacon_proposer.rs
@@ -4,7 +4,8 @@ use actix_web::{
     HttpResponse, Responder, post,
     web::{Data, Json},
 };
-use ream_api_types_beacon::{error::ApiError, request::PrepareBeaconProposerItem};
+use ream_api_types_beacon::request::PrepareBeaconProposerItem;
+use ream_api_types_common::error::ApiError;
 use ream_fork_choice::store::Store;
 use ream_operation_pool::OperationPool;
 use ream_storage::db::ReamDB;
