+++ crates/rpc/beacon/src/handlers/state.rs
@@ -4,11 +4,10 @@ use actix_web::{
 };
 use alloy_primitives::B256;
 use ream_api_types_beacon::{
-    error::ApiError,
     query::EpochQuery,
     responses::{BeaconResponse, BeaconVersionedResponse},
 };
-use ream_api_types_common::id::ID;
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_consensus_beacon::electra::beacon_state::BeaconState;
 use ream_consensus_misc::{
     checkpoint::Checkpoint, constants::beacon::SYNC_COMMITTEE_SIZE,
