+++ crates/rpc/beacon/src/handlers/validator.rs
@@ -5,14 +5,13 @@ use actix_web::{
     web::{Data, Json, Path, Query},
 };
 use ream_api_types_beacon::{
-    error::ApiError,
     id::ValidatorID,
     query::{AttestationQuery, IdQuery, StatusQuery},
     request::ValidatorsPostRequest,
     responses::{BeaconResponse, DataResponse},
     validator::{ValidatorBalance, ValidatorData, ValidatorStatus},
 };
-use ream_api_types_common::id::ID;
+use ream_api_types_common::{error::ApiError, id::ID};
 use ream_bls::PublicKey;
 use ream_consensus_beacon::electra::beacon_state::BeaconState;
 use ream_consensus_misc::{
