+++ crates/rpc/lean/src/handlers/peer.rs
@@ -3,7 +3,7 @@ use std::{collections::HashMap, sync::Arc};
 use actix_web::{HttpResponse, Responder, get, web::Data};
 use libp2p::PeerId;
 use parking_lot::Mutex;
-use ream_api_types_beacon::error::ApiError;
+use ream_api_types_common::error::ApiError;
 use ream_p2p::network::peer::ConnectionState;
 use ream_rpc_beacon::handlers::peers::PeerCount;
